Revert "Feature: [STORE] Refactor legacy stores into main store with slices"
