+++ apps/editor/src/App.tsx
@@ -1,3 +1,4 @@
+import { useEditorStore } from "./stores/editorStore";
 import ResizablePanel from "./components/ResizablePanel";
 import SceneTreeTabs from "./components/sceneTree/SceneTreeTabs";
 import InspectorTabs from "./components/inspector/InspectorTabs";
@@ -6,10 +7,9 @@ import { Folder } from "lucide-react";
 import Icon from "./components/Icon";
 // import SaveSceneButton from "./components/SaveSceneButton";
 import SceneEditor from "./components/sceneEditor";
-import { useEditor } from "./hooks/useEditor";
 
 function App() {
-    const { panelSizes, setPanelSize } = useEditor();
+    const { panelSizes, setPanelSize } = useEditorStore();
 
     return (
         <div className="h-screen bg-background-primary flex flex-col relative">
