--- apps/editor/src/stores/slices/index.ts
@@ -1,11 +1,0 @@
-// ============================================================================
-// SLICES INDEX - Export all slice interfaces and creators
-// ============================================================================
-
-export type { EditorSlice } from './editorSlice';
-export type { SceneSlice } from './sceneSlice';
-export type { UISlice } from './uiSlice';
-
-export { createEditorSlice } from './editorSlice';
-export { createSceneSlice } from './sceneSlice';
-export { createUISlice } from './uiSlice';
